refactor: 전역 및 페이지 전용 컴포넌트 디렉터리 구조화 (#11)
