+++ src/app/components/progress-section.tsx
