+++ src/app/components/survey-preview-section.tsx
@@ -1,9 +1,9 @@
-import { SurveyCard } from '@/components/survey/survey-card'
-import { SectionHeader } from './section-header'
+import { SurveyCard } from '@/app/components/survey-card'
+import { SectionHeader } from '@/app/components/section-header'
 import { ResponsiveSliderGrid } from '@/components/ui/responsive-slider-grid'
 import { SurveyItem } from '@/types/survey'
 
-export function WaitingSurveySection({
+export function SurveyPreviewSection({
   items,
 }: {
   items: (SurveyItem & { answers: string[] })[]
@@ -11,9 +11,8 @@ export function WaitingSurveySection({
   return (
     <section>
       <SectionHeader
-        title="응답자를 기다리고 있어요!"
-        subtitle="기간이 설정됐지만 아직 응답자가 충분하지 않은 설문이에요"
-        moreHref="/survey?tab=waiting"
+        title="이런 설문조사는 어때요?"
+        moreHref="/all-survey"
       />
       <ResponsiveSliderGrid>
         {items.map((s) => (
@@ -24,9 +23,9 @@ export function WaitingSurveySection({
             thumbnail={s.thumbnail}
             reward={s.reward}
             duration={s.duration}
-            tags={s.tags}
             variant="withAnswers"
             answers={s.answers}
+            tags={s.tags}
           />
         ))}
       </ResponsiveSliderGrid>
