+++ src/app/components/waiting-survey-section.tsx
@@ -1,9 +1,9 @@
-import { SurveyCard } from '@/components/survey/survey-card'
-import { SectionHeader } from './section-header'
+import { SurveyCard } from '@/app/components/survey-card'
+import { SectionHeader } from '@/app/components/section-header'
 import { ResponsiveSliderGrid } from '@/components/ui/responsive-slider-grid'
 import { SurveyItem } from '@/types/survey'
 
-export function NewestSurveySection({
+export function WaitingSurveySection({
   items,
 }: {
   items: (SurveyItem & { answers: string[] })[]
@@ -11,8 +11,9 @@ export function NewestSurveySection({
   return (
     <section>
       <SectionHeader
-        title="방금 올라온 따끈따끈한 설문조사"
-        moreHref="/survey?tab=newest"
+        title="응답자를 기다리고 있어요!"
+        subtitle="기간이 설정됐지만 아직 응답자가 충분하지 않은 설문이에요"
+        moreHref="/all-survey?tab=waiting"
       />
       <ResponsiveSliderGrid>
         {items.map((s) => (
