+++ src/app/page.tsx
@@ -1,11 +1,11 @@
-import { AdSection, BannerItem } from '@/components/home/ad-section'
-import { StatusSection } from '@/components/home/status-section'
-import { CustomSurveySection } from '@/components/home/custom-survey-section'
-import { UrgentCarouselSection } from '@/components/home/urgent-carousel-section'
-import { SurveyPreviewSection } from '@/components/home/survey-preview-section'
-import { WaitingSurveySection } from '@/components/home/waiting-survey-section'
-import { NewestSurveySection } from '@/components/home/newest-survey-section'
-import { BestRewardProductsSection } from '@/components/home/best-reward-products-section'
+import { AdSection, BannerItem } from '@/app/components/ad-section'
+import { StatusSection } from '@/app/components/status-section'
+import { CustomSurveySection } from '@/app/components/custom-survey-section'
+import { UrgentCarouselSection } from '@/app/components/urgent-carousel-section'
+import { SurveyPreviewSection } from '@/app/components/survey-preview-section'
+import { WaitingSurveySection } from '@/app/components/waiting-survey-section'
+import { NewestSurveySection } from '@/app/components/newest-survey-section'
+import { BestRewardProductsSection } from '@/app/components/best-reward-products-section'
 import { SurveyItem } from '@/types/survey'
 import { ProductItem } from '@/types/product'
