+++ src/components/layout/header.tsx
@@ -47,7 +47,7 @@ export function Header() {
         {/* 중간: 1차 내비게이션 */}
         <nav className="hidden md:flex items-center gap-8">
           <NavLink
-            href="/survey"
+            href="/all-survey"
             label="설문조사"
           />
           <NavLink
